Setup e2e harness & first e2e test:
+ use nightwatchjs and selenium to automate browser (casperjs won't work for us, more below)
+ tweak server.babel.js to load alternate .babelrc for test env
+ add e2e script to bootstrap e2e tests and manage child processes
+ add process management helper:
  + killOnExit:
    /**
     * Ensures `processesToKill` and their child processes are killed
     * when `exitingProcess` exits, receives `SIGINT`, or has an
     * uncaught exception
     * @param {ChildProcess} exitingProcess - process to watch
     * @param {Array.<ChildProcess>} processesToKill - processes to kill
     */
+ stub out process logging framework (see scripts/helpers/logger/*):
  + add selenium logging to file
+ wip - working signup-e2e test
+ comment on possible future optimization using web pack, caching, and watchers

Testing with ES6
=============

Because we want to use ES6 in our tests (to be consistent within the codebase), we must transpile the test code. The simplest way to do this is by using mocha's `--compilers` option; however, that option is only accessible via CLI, not programatically. Therefore, we must call the `mocha` binary as the starting point for running our test suite.

In the world of E2E tests, everyone wants you to use their binary (casperjs, phantomjs, nightwatchjs, etc.) which conflicts with our constraint. After trying many combinations and permutations the best option for starting from the mocha binary to run e2e tests relies on using nightwatchjs's node api, which is supported and [documented](http://nightwatchjs.org/guide#using-the-standard-mocha).